feat: add check for special characters in hcp name, move validation logic into separate file
