+++ src/services/submissionService.ts
@@ -3,9 +3,11 @@ import * as gqlTypes from '../typeDefs/gqlTypes.js'
 import * as dbSchema from '../typeDefs/dbSchema.js'
 import { dbInstance } from '../firebaseDb.js'
 import { ErrorCode, Result } from '../result.js'
-import { hasSpecialCharacters, hasScriptTags } from '../../utils/stringUtils.js'
+import { hasSpecialCharacters } from '../../utils/stringUtils.js'
 import { createFacility } from './facilityService.js'
 import { createHealthcareProfessional } from './healthcareProfessionalService.js'
+import { validateSubmissionSearchFilters } from '../validation/validateSubmissions.js'
+
 
 /**
  * Gets the Submission from the database that matches the id.
@@ -384,266 +386,6 @@ function validateIdInput(id: string): Result<unknown> {
     }
 
     return validationResults
-}
-
-function validateSubmissionOrderBy(filters: gqlTypes.SubmissionSearchFilters): Result<unknown> {
-    const validateOrderByResults: Result<unknown> = {
-        data: undefined,
-        hasErrors: false,
-        errors: []
-    }
-
-    if (!filters.orderBy || filters.orderBy.length === 0) {
-        return validateOrderByResults
-    }
-    
-    const validOrderByFields = [
-        'googleMapsUrl',
-        'healthcareProfessionalName',
-        'isUnderReview',
-        'isApproved',
-        'isRejected',
-        'createdDate',
-        'updatedDate',
-    ]
-
-    if (filters.orderBy?.some(order => order.fieldToOrder === 'id')) {
-        validateOrderByResults.hasErrors = true
-        validateOrderByResults.errors?.push({
-            field: 'orderBy',
-            errorCode: ErrorCode.ORDERBY_OPTION_INVALID,
-            httpStatus: 400
-        })
-    }
-
-    const invalidOrderByOption = filters.orderBy?.find(order => 
-        !validOrderByFields.includes(order.fieldToOrder))
-
-    if (invalidOrderByOption) {
-        validateOrderByResults.hasErrors = true
-        validateOrderByResults.errors?.push({
-            field: 'orderBy',
-            errorCode: ErrorCode.ORDERBY_OPTION_NOT_FOUND,
-            httpStatus: 400
-        })
-    }
-
-    if (filters.orderBy && filters.orderBy.length > 2) {
-        validateOrderByResults.hasErrors = true
-        validateOrderByResults.errors?.push({
-            field: 'orderBy',
-            errorCode: ErrorCode.ORDERBY_OPTION_MAX_LIMIT,
-            httpStatus: 400
-        })
-    }
-
-    if (filters.orderBy && filters.orderBy.length === 2 && 
-        filters.orderBy[0].fieldToOrder === filters.orderBy[1].fieldToOrder) {
-        validateOrderByResults.hasErrors = true
-        validateOrderByResults.errors?.push({
-            field: 'orderBy',
-            errorCode: ErrorCode.ORDERBY_FIELD_SELECTED_TWICE,
-            httpStatus: 400
-        })
-    }
-
-    return validateOrderByResults
-}
-
-function validateSubmissionSearchBySpokenLanguage(filters: gqlTypes.SubmissionSearchFilters): Result<unknown> {
-    const validateSpokenLanguageResults: Result<unknown> = {
-        data: undefined,
-        hasErrors: false,
-        errors: []
-    }
-
-    if (!filters.spokenLanguages) {
-        return validateSpokenLanguageResults
-    }
-
-    if (filters.spokenLanguages.length == 0) {
-        validateSpokenLanguageResults.hasErrors = true
-        validateSpokenLanguageResults.errors?.push({
-            field: 'spokenLanguages',
-            errorCode: ErrorCode.MISSING_INPUT,
-            httpStatus: 400
-        })
-    }
-    
-    for (const lang of filters.spokenLanguages) {
-        if (!Object.values(gqlTypes.Locale).includes(lang)) {
-            validateSpokenLanguageResults.hasErrors = true
-            validateSpokenLanguageResults.errors?.push({
-                field: 'spokenLanguages',
-                errorCode: ErrorCode.INVALID_LANGUAGE_INPUT,
-                httpStatus: 400
-            })
-        }
-    }
-    
-    const languageCodeOccurrences: any = {}
-    for (let lang of filters.spokenLanguages) {
-        if(languageCodeOccurrences[lang]) {     
-            validateSpokenLanguageResults.hasErrors = true
-            validateSpokenLanguageResults.errors?.push({
-                field: 'spokenLanguages',
-                errorCode: ErrorCode.INVALID_LANGUAGE_INPUT,
-                httpStatus: 400
-            })
-            break
-        } else {
-            languageCodeOccurrences[lang] = 1
-        }
-    }
-
-    return validateSpokenLanguageResults
-}
-
-function validateSubmissionSearchFilters(filters: gqlTypes.SubmissionSearchFilters): Result<unknown> {
-    const validateSearchResults: Result<unknown> = {
-        data: undefined,
-        hasErrors: false,
-        errors: []
-    }
-
-    const orderByValidationResult = validateSubmissionOrderBy(filters)
-    if(orderByValidationResult.hasErrors) {
-        validateSearchResults.hasErrors = true
-        validateSearchResults.errors = (validateSearchResults.errors || []).concat(orderByValidationResult.errors || [])
-    }
-
-    const spokenLanguageResult = validateSubmissionSearchBySpokenLanguage(filters)
-    if (spokenLanguageResult.hasErrors) {
-        validateSearchResults.hasErrors = true
-        validateSearchResults.errors = (validateSearchResults.errors || []).concat(spokenLanguageResult.errors || [])
-    }
-
-    if (filters.limit && filters.limit > 1000) {
-        validateSearchResults.hasErrors = true
-        validateSearchResults.errors?.push({
-            field: 'limit',
-            errorCode: ErrorCode.MAX_LIMIT,
-            httpStatus: 400
-        })
-    }
-
-    if (filters.limit && filters.limit < 0) {
-        validateSearchResults.hasErrors = true
-        validateSearchResults.errors?.push({
-            field: 'limit',
-            errorCode: ErrorCode.MIN_LIMIT,
-            httpStatus: 400
-        })
-    }
-
-    if (filters.offset && filters.offset > 10000) {
-        validateSearchResults.hasErrors = true
-        validateSearchResults.errors?.push({
-            field: 'offset',
-            errorCode: ErrorCode.MAX_LIMIT,
-            httpStatus: 400
-        })
-    }
-
-    if (filters.offset && filters.offset < 0) {
-        validateSearchResults.hasErrors = true
-        validateSearchResults.errors?.push({
-            field: 'offset',
-            errorCode: ErrorCode.MIN_LIMIT,
-            httpStatus: 400
-        })
-    }
-
-    if (filters.googleMapsUrl === '' || (filters.googleMapsUrl && !filters.googleMapsUrl.trim())) {
-        validateSearchResults.hasErrors = true
-        validateSearchResults.errors?.push({
-            field: 'googleMapsUrl',
-            errorCode: ErrorCode.MISSING_INPUT,
-            httpStatus: 400
-        })
-    }
-
-    if (filters.googleMapsUrl && filters.googleMapsUrl.length > 1028) {
-        validateSearchResults.hasErrors = true
-        validateSearchResults.errors?.push({
-            field: 'googleMapsUrl',
-            errorCode: ErrorCode.INVALID_LENGTH_TOO_LONG,
-            httpStatus: 400
-        })
-    }
-
-    if (filters.googleMapsUrl && hasScriptTags(filters.googleMapsUrl)) {
-        validateSearchResults.hasErrors = true
-        validateSearchResults.errors?.push({
-            field: 'googleMapsUrl',
-            errorCode: ErrorCode.INVALID_INPUT,
-            httpStatus: 400
-        })
-    }
-
-    if (filters.healthcareProfessionalName === '' || (filters.healthcareProfessionalName && !filters.healthcareProfessionalName.trim())) {
-        validateSearchResults.hasErrors = true
-        validateSearchResults.errors?.push({
-            field: 'healthcareProfessionalName',
-            errorCode: ErrorCode.MISSING_INPUT,
-            httpStatus: 400
-        })
-    }
-
-    if (filters.healthcareProfessionalName && filters.healthcareProfessionalName.length > 128) {
-        validateSearchResults.hasErrors = true
-        validateSearchResults.errors?.push({
-            field: 'healthcareProfessionalName',
-            errorCode: ErrorCode.INVALID_LENGTH_TOO_LONG,
-            httpStatus: 400
-        })
-    }
-
-    if (filters.healthcareProfessionalName && hasScriptTags(filters.healthcareProfessionalName)) {
-        validateSearchResults.hasErrors = true
-        validateSearchResults.errors?.push({
-            field: 'healthcareProfessionalName',
-            errorCode: ErrorCode.INVALID_INPUT,
-            httpStatus: 400
-        })
-    }
-
-    // if (filters.spokenLanguages) {
-    //     if (filters.spokenLanguages.length === 0) {
-    //         validateSearchResults.hasErrors = true
-    //         validateSearchResults.errors?.push({
-    //             field: 'spokenLanguages',
-    //             errorCode: ErrorCode.MISSING_INPUT,
-    //             httpStatus: 400
-    //         })
-    //     } else {
-    //         for (const lang of filters.spokenLanguages) {
-    //             if (!Object.values(gqlTypes.Locale).includes(lang)) {
-    //                 validateSearchResults.hasErrors = true
-    //                 validateSearchResults.errors?.push({
-    //                     field: 'spokenLanguages',
-    //                     errorCode: ErrorCode.INVALID_LANGUAGE_INPUT,
-    //                     httpStatus: 400
-    //                 })
-    //             }
-    //         }
-    //     }
-    // }
-
-    // if (filters.spokenLanguages && filters.spokenLanguages.length >= 2) {
-    //     const languageCodeOccurances: any = {}
-    //     for (let lang of filters.spokenLanguages) {
-    //         !languageCodeOccurances[lang] ? languageCodeOccurances[lang] = 1 : 
-    //         validateSearchResults.hasErrors = true
-    //         validateSearchResults.errors?.push({
-    //             field: 'spokenLanguages',
-    //             errorCode: ErrorCode.INVALID_LANGUAGE_INPUT,
-    //             httpStatus: 400
-    //         })
-    //     }
-    // }
-
-    return validateSearchResults
 }
 
 const validateSubmissionInputFields = (input: gqlTypes.CreateSubmissionInput): Result<unknown> => {
